+++ bioshell-ff/src/system.rs
@@ -20,14 +20,14 @@ impl System {
     pub fn coordinates(&self) -> &Coordinates { &self.coordinates }
 
     /// Current volume of the simulation box
-    pub fn volume(&self) -> f32 { self.coordinates.box_len().powf(3.0) }
+    pub fn volume(&self) -> f64 { self.coordinates.box_len().powf(3.0) }
 
     /// Returns the simulation box length
     #[inline(always)]
-    pub fn box_len(&self) -> f32 { self.coordinates.box_len() }
+    pub fn box_len(&self) -> f64 { self.coordinates.box_len() }
 
     /// Changes the simulation box length which results in the change of all positions and the volume
-    pub fn set_box_len(&mut self, new_box_len: f32) {
+    pub fn set_box_len(&mut self, new_box_len: f64) {
 
         // ---------- expansion / contraction factor
         let f = new_box_len/self.coordinates().box_len();
@@ -48,12 +48,12 @@ impl System {
     /// Returns the number of atoms of this system
     pub fn size(&self) -> usize { self.coordinates.size() }
 
-    pub fn set(&mut self, i:usize, x: f32, y: f32, z: f32) {
+    pub fn set(&mut self, i:usize, x: f64, y: f64, z: f64) {
         self.coordinates.set(i, x, y, z);
         self.neighbor_list.update(&self.coordinates, i);
     }
 
-    pub fn add(&mut self, i:usize, x: f32, y: f32, z: f32) {
+    pub fn add(&mut self, i:usize, x: f64, y: f64, z: f64) {
         self.coordinates.add(i, x, y, z);
         self.neighbor_list.update(&self.coordinates, i);
     }
@@ -65,16 +65,16 @@ impl System {
     }
 
     /// Provides the interaction cutoff radius used by the neighbor list
-    pub fn cutoff(&self) -> f32 { self.neighbor_list.cutoff() }
+    pub fn cutoff(&self) -> f64 { self.neighbor_list.cutoff() }
 
     /// Modifies the interaction cutoff radius used by the neighbor list
-    pub fn set_cutoff(&mut self, d0: f32) { self.neighbor_list.set_cutoff(d0); }
+    pub fn set_cutoff(&mut self, d0: f64) { self.neighbor_list.set_cutoff(d0); }
 
     /// Provides the width of the buffer zone used by the neighbor list
-    pub fn buffer_width(&self) -> f32 { self.neighbor_list.buffer_width() }
+    pub fn buffer_width(&self) -> f64 { self.neighbor_list.buffer_width() }
 
     /// Modifies the width of the buffer zone used by the neighbor list
-    pub fn set_buffer_width(&mut self, width: f32) { self.neighbor_list.set_buffer_width(width); }
+    pub fn set_buffer_width(&mut self, width: f64) { self.neighbor_list.set_buffer_width(width); }
 
     /// Sets new rules that define which atoms and atom pairs can be neighbors.
     /// Note: remember to update this neighbor list after this call!
