+++ bioshell-sim/src/generators.rs
@@ -5,24 +5,24 @@ use bioshell_numerical::Vec3;
 
 pub fn random_chain(bond_length:f64, repulsion_distance: f64, start: &Vec3, system: &mut Coordinates) {
 
-    let cutoff2 = (repulsion_distance * repulsion_distance) as f32;
+    let cutoff2 = (repulsion_distance * repulsion_distance) as f64;
 
     system.set_x(0, start.x);
     system.set_y(0, start.y);
     system.set_z(0, start.z);
 
     let (x, y, z) = random_unit_versor();
-    system.set_x(1, system.x(0) + (x * bond_length) as f32);
-    system.set_y(1, system.y(0) + (y * bond_length) as f32);
-    system.set_z(1, system.z(0) + (z * bond_length) as f32);
+    system.set_x(1, system.x(0) + (x * bond_length) as f64);
+    system.set_y(1, system.y(0) + (y * bond_length) as f64);
+    system.set_z(1, system.z(0) + (z * bond_length) as f64);
 
     for i in 2..system.size() {
         let mut go_on: bool = true;
         while go_on {
             let (x, y, z) = random_unit_versor();
-            system.set_x(i, system.x(i-1) + (x * bond_length) as f32);
-            system.set_y(i, system.y(i-1) + (y * bond_length) as f32);
-            system.set_z(i, system.z(i-1) + (z * bond_length) as f32);
+            system.set_x(i, system.x(i-1) + (x * bond_length) as f64);
+            system.set_y(i, system.y(i-1) + (y * bond_length) as f64);
+            system.set_z(i, system.z(i-1) + (z * bond_length) as f64);
             go_on = false;
             for j in 0..(i-1) {
                 if system.closest_distance_square(j, i) <= cutoff2 {
@@ -37,14 +37,14 @@ pub fn random_chain(bond_length:f64, repulsion_distance: f64, start: &Vec3, syst
 pub fn cubic_grid_atoms(system: &mut Coordinates) {
 
     let points_one_side: usize = (f64::powf(system.size() as f64, 1.0 / 3.0)).ceil() as usize;
-    let dw = system.box_len() / points_one_side as f32;
+    let dw = system.box_len() / points_one_side as f64;
     let cell_margin = dw / 2.0;
 
     for i in 0..system.size() {
         let k = i % points_one_side;
         let l = (i / points_one_side) % points_one_side;
         let m = (i / (points_one_side * points_one_side)) % points_one_side;
-        system.set(i,dw * k as f32 + cell_margin,dw * l as f32 + cell_margin,dw * m as f32 + cell_margin)
+        system.set(i,dw * k as f64 + cell_margin,dw * l as f64 + cell_margin,dw * m as f64 + cell_margin)
     }
 }
 
@@ -52,13 +52,13 @@ pub fn cubic_grid_atoms(system: &mut Coordinates) {
 pub fn square_grid_atoms(system: &mut Coordinates) {
 
     let points_one_side: usize = (f64::powf(system.size() as f64, 0.5)).ceil() as usize;
-    let dw = system.box_len() / points_one_side as f32;
+    let dw = system.box_len() / points_one_side as f64;
     let cell_margin = dw / 2.0;
 
     for i in 0..system.size() {
         let k = i % points_one_side;
         let l = i / points_one_side;
-        system.set(i,dw * k as f32 + cell_margin,dw * l as f32 + cell_margin,0.0);
+        system.set(i,dw * k as f64 + cell_margin,dw * l as f64 + cell_margin,0.0);
     }
 }
