+++ src/polymer.rs
@@ -26,16 +26,16 @@ impl NbListRules for PolymerRules {
 }
 
 pub fn main() {
-    const E_REP: f32 = 4.25;
-    const E_FROM: f32 = 4.5;
-    const E_TO: f32 = 6.0;
-    const E_VAL: f32 = -1.0;
-    const REP_VAL: f32 = 1000.0;
+    const E_REP: f64 = 4.25;
+    const E_FROM: f64 = 4.5;
+    const E_TO: f64 = 6.0;
+    const E_VAL: f64 = -1.0;
+    const REP_VAL: f64 = 1000.0;
     const N_BEADS: usize = 300;
     const N_SMALL_CYCLES: i32 = 1000;
     const N_LARGE_CYCLES: i16 = 10000;
     const T: f64 = 1.8;
-    const L: f32 = 900.0;
+    const L: f64 = 900.0;
 
     // ---------- Create system's coordinates
     let mut coords = Coordinates::new(N_BEADS);
@@ -50,7 +50,7 @@ pub fn main() {
     to_pdb(&system.coordinates(), 0, "tra.pdb");
 
     // ---------- Contact energy
-    let contacts = PairwiseNonbondedEvaluator::new(E_TO as f32,
+    let contacts = PairwiseNonbondedEvaluator::new(E_TO as f64,
             Box::new(SimpleContact::new(E_REP,E_FROM,E_TO,REP_VAL,E_VAL)));
     // ---------- Harmonic energy (i.e. springs between beads)
     let harmonic = SimpleHarmonic::new(3.8,5.0);
@@ -62,9 +62,9 @@ pub fn main() {
 
     let mut sampler = IsothermalMC::new(T, Ensemle::NVT, 1.0);
     sampler.energy = Box::new(total);    // --- The total has been moved to a box within the sampler
-    let m: Box<dyn Fn(&mut System,f32) -> Range<usize>> = Box::new(single_atom_move);
+    let m: Box<dyn Fn(&mut System,f64) -> Range<usize>> = Box::new(single_atom_move);
     sampler.add_mover(m,3.0);
-    // let m: Box<dyn Fn(&mut Coordinates,f32) -> Range<usize>> = Box::new(perturb_chain_fragment);
+    // let m: Box<dyn Fn(&mut Coordinates,f64) -> Range<usize>> = Box::new(perturb_chain_fragment);
     // sampler.add_mover(m,5.0);
 
     let start = Instant::now();
